# conditionally use `SpecialFunctions` through weak deps or `Requires`