--- a/Project.toml
+++ b/Project.toml
@@ -6,6 +6,7 @@ version = "0.9.9"
 ColorTypes = "3da002f7-5984-5a60-b8a6-cbb66c0b333f"
 FixedPointNumbers = "53c48c17-4a7d-5ca2-90c5-79b7896eea93"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 TensorCore = "62fd8b95-f654-4bbd-a8a5-9c27f68ccd50"
 
@@ -18,6 +19,7 @@ SpecialFunctionsExt = "SpecialFunctions"
 [compat]
 ColorTypes = "0.10, 0.11"
 FixedPointNumbers = "0.8.2"
+Requires = "1"
 TensorCore = "0.1"
 julia = "1"