--- a/Project.toml
+++ b/Project.toml
@@ -10,18 +10,19 @@ Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 TensorCore = "62fd8b95-f654-4bbd-a8a5-9c27f68ccd50"
 
+[compat]
+ColorTypes = "0.10, 0.11"
+FixedPointNumbers = "0.8.2"
+Requires = "1"
+SpecialFunctions = "0.8, 0.9, 0.10, 1, 2.0"
+TensorCore = "0.1"
+julia = "1"
+
 [weakdeps]
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
 
 [extensions]
 SpecialFunctionsExt = "SpecialFunctions"
-
-[compat]
-ColorTypes = "0.10, 0.11"
-FixedPointNumbers = "0.8.2"
-Requires = "1"
-TensorCore = "0.1"
-julia = "1"
 
 [extras]
 Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"