--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -280,9 +280,14 @@ ColorTypes.comp2(c::RGBA32) = alpha(c)
         for g in (Gray(0.4), Gray{N0f8}(0.4))
             @test @inferred(zero(g)) === typeof(g)(0)
             @test @inferred(oneunit(g)) === typeof(g)(1)
-            for opgroup in (ColorVectorSpace.unaryOps, (:trunc, :floor, :round, :ceil, :eps, :bswap))
-                for op in opgroup
-                    op ∈ (:frexp, :exponent, :modf, :lfact) && continue
+            for mod in (
+                ColorVectorSpace,
+                Base.get_extension(ColorVectorSpace, :SpecialFunctionsExt),
+                (; UnaryOps = (:trunc, :floor, :round, :ceil, :eps, :bswap)),
+            )
+                mod === nothing && continue  # pre weak dependencies
+                for op in mod.UnaryOps
+                    op ∈ (:frexp, :exponent, :modf, :logfactorial) && continue
                     op === :~ && eltype(g) === Float64 && continue
                     op === :significand && eltype(g) === N0f8 && continue
                     try