Fix https://github.com/JuliaGraphics/ColorVectorSpace.jl/issues/184.
Fix https://github.com/JuliaGraphics/ColorVectorSpace.jl/issues/182.
Fix last part of https://github.com/JuliaPlots/UnicodePlots.jl/issues/291.

This is a proposal to make all the `SpecialFunctions` related functions (which are clearly piracy) optional using weak dependencies on `julia` >= 1.9 and `Requires` on older releases.

This would be (slightly) breaking so I guess we should bump `ColorVectorSpace` to `0.10.0` if this PR is merged, since you'd have to use:
```julia
julia> using SpecialFunctions
julia> using ColorVectorSpace
```
to recover all the functionalities on current master.

Note that `lfact` is renamed to `logfactorial` since `lfact` was deprecated in `SpecialFunctions`.

I followed the guidelines in https://pkgdocs.julialang.org/dev/creating-packages/#Conditional-loading-of-code-in-packages-(Extensions) and https://docs.julialang.org/en/v1.10-dev/manual/code-loading/#man-extensions.

This PR was written for `ColorSchemes`, and `pkg> test ColorSchemes` passes locally since `ColorSchemes` does not depend on the `SpecialFunctions` functionalities defined in `ColorVectorSpace`.

Latency improvements:
<details>

**PR**
```julia
julia> @time @time_imports using ColorVectorSpace
      1.0 ms  Statistics
     66.4 ms  FixedPointNumbers
    100.7 ms  ColorTypes
      0.4 ms  TensorCore
    319.3 ms  ColorVectorSpace
  0.613890 seconds (553.44 k allocations: 39.516 MiB, 17.56% gc time, 16.28% compilation time)
```

**master**
```julia
julia> @time @time_imports using ColorVectorSpace
      1.5 ms  Statistics
     85.7 ms  FixedPointNumbers
    124.7 ms  ColorTypes
      8.0 ms  IrrationalConstants
      0.2 ms  Compat
    105.7 ms  ChainRulesCore
      1.4 ms  DocStringExtensions
      0.4 ms  LogExpFunctions
      0.3 ms  OpenLibm_jll
     23.2 ms  Preferences
      0.3 ms  JLLWrappers
      7.0 ms  OpenSpecFun_jll 90.69% compilation time
     22.2 ms  SpecialFunctions
      0.3 ms  TensorCore
    203.7 ms  ColorVectorSpace
  0.642129 seconds (628.22 k allocations: 43.283 MiB, 5.46% compilation time)
```

</details>

Ping @timholy or @kimikage or @johnnychen94 for review (primary contributors to this repo).